Unsucessfully made shmsize into string

+++ master/pkg/tasks/task_trial.go
@@ -121,7 +121,7 @@ func (s TrialSpec) ToTaskSpec(keys *ssh.PrivateAndPublicKeys) TaskSpec {
 	}
 
 	if shm := s.ExperimentConfig.Resources().ShmSize(); shm != nil {
-		res.ShmSize = int64(*shm)
+		res.ShmSize = *shm
 	}
 
 	mounts := ToDockerMounts(s.ExperimentConfig.BindMounts(), res.WorkDir)
